pyproject.toml
[tool.poetry]
name = "gd-pm"
version = "0.1.0"
description = ""
authors = ["Frank Zhang <frankzhang123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pycryptodome = "^3.20.0"
python-dotenv = "^1.0.1"
lark-oapi = "^1.2.8"
sqlite-utils = "^3.36"
dash = "^2.17.1"
attr = "^0.3.2"
crypto = "^1.4.1"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
